drivers: serial: Add CONFIG_CDCACM_NON_BLOCKING
To mimic the behavior of a usb serial converter, which does not care
whether there is a connection or not and does not block writing to the
buffer. ⚠️ WARNING ⚠️: this was developed for NSH access to help debug
by printf, if you have to make sure that what you send is what will be
transmitted don't use this!

Signed-off-by: Matheus Castello <matheus@castello.eng.br>